Implement changes requested in review
